Added svg saving
