+++ .gitignore
@@ -18,6 +18,8 @@ mono_crash.*
 
 # Visual Studio 2015/2017 cache/options directory
 .vs/
+bin/
+obj/Debug/
 # Uncomment if you have tasks that create the project's static files in wwwroot
 #wwwroot/
