+++ Weave.sln
@@ -1,9 +1,9 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.8.34322.80
+VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Weave", "Weave\Weave.csproj", "{6999CE64-AD0F-4F7B-A3F4-70D2701522E8}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Weave", "Weave.csproj", "{E662815C-E7A9-4493-A673-6799A8962282}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,15 +11,15 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{6999CE64-AD0F-4F7B-A3F4-70D2701522E8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6999CE64-AD0F-4F7B-A3F4-70D2701522E8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6999CE64-AD0F-4F7B-A3F4-70D2701522E8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6999CE64-AD0F-4F7B-A3F4-70D2701522E8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E662815C-E7A9-4493-A673-6799A8962282}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E662815C-E7A9-4493-A673-6799A8962282}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E662815C-E7A9-4493-A673-6799A8962282}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E662815C-E7A9-4493-A673-6799A8962282}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {350C0D71-5A3B-472D-B41A-7E667F854526}
+		SolutionGuid = {BF8F7496-93F4-4364-BE80-197446E7FF46}
 	EndGlobalSection
 EndGlobal
