--- Weave/bin/Debug/net8.0/Weave.deps.json
@@ -1,23 +1,0 @@
-{
-  "runtimeTarget": {
-    "name": ".NETCoreApp,Version=v8.0",
-    "signature": ""
-  },
-  "compilationOptions": {},
-  "targets": {
-    ".NETCoreApp,Version=v8.0": {
-      "Weave/1.0.0": {
-        "runtime": {
-          "Weave.dll": {}
-        }
-      }
-    }
-  },
-  "libraries": {
-    "Weave/1.0.0": {
-      "type": "project",
-      "serviceable": false,
-      "sha512": ""
-    }
-  }
-}
